Move Bid type to types directory
