+++ packages/frontend/src/providers/BidsProvider/index.ts
@@ -1,2 +1,1 @@
 export { useBids, BidsProvider } from './provider'
-export type { Bid } from './types'
